# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Point Market API",
        "description": "Dokumentasi API untuk seluruh fitur aplikasi Point Market",
        "contact": {
            "email": "admin@pointmarket.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/riwayat_misi": {
            "get": {
                "tags": [
                    "Riwayat Misi"
                ],
                "summary": "Get all Riwayat Misi",
                "description": "Retrieve a list of all Riwayat Misi",
                "operationId": "getAllRiwayatMisi",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_misi": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "npm": {
                                                "type": "string",
                                                "example": "12345678"
                                            },
                                            "nama_misi": {
                                                "type": "string",
                                                "example": "Seminar AI"
                                            },
                                            "deskripsi": {
                                                "type": "string",
                                                "example": "Mengikuti seminar tentang AI"
                                            },
                                            "point": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "tanggal_transaksi": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-01-22"
                                            },
                                            "file_bukti": {
                                                "type": "string",
                                                "example": "bukti_seminar.pdf"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pending"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Riwayat Misi"
                ],
                "summary": "Create a new Riwayat Misi",
                "description": "Store a new Riwayat Misi to the database",
                "operationId": "createRiwayatMisi",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "npm": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "nama_misi": {
                                        "type": "string",
                                        "example": "Seminar AI"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Mengikuti seminar tentang AI"
                                    },
                                    "point": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "file_bukti": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "file.pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Riwayat Misi created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Misi berhasil diajukan!"
                                        },
                                        "data": {
                                            "properties": {
                                                "id_misi": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "npm": {
                                                    "type": "string",
                                                    "example": "12345678"
                                                },
                                                "nama_misi": {
                                                    "type": "string",
                                                    "example": "Seminar AI"
                                                },
                                                "deskripsi": {
                                                    "type": "string",
                                                    "example": "Mengikuti seminar tentang AI"
                                                },
                                                "point": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "tanggal_transaksi": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-01-22"
                                                },
                                                "file_bukti": {
                                                    "type": "string",
                                                    "example": "uploads/file.pdf"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}